+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
 ## [Unreleased]
+- in `check-chef-nodes.rb` made client name required
+- in `check-chef-nodes.rb` made a sane default for client name
 
 ## [2.0.0] - 2016-07-25
 ### Changed